// lib/types.ts
export interface Fault {
  code: string;
  message: string;
  detail?:
    | FaultStruct
    | {
        name: string;
        message: string;
        stack?: string;
      };
  timestamp?: number;
}

export interface SessionFault extends Fault {
  timestamp: number;
  provisions: [string, ...Value[]][];
  retryNow?: boolean;
  precondition?: boolean;
  retries: number;
  expiry?: number;
}

export interface Attributes {
  object?: [number, 1 | 0];
  writable?: [number, 1 | 0];
  value?: [number, [string | number | boolean, string]];
  notification?: [number, number];
  accessList?: [number, string[]];
}

export interface AttributeTimestamps {
  object?: number;
  writable?: number;
  value?: number;
  notification?: number;
  accessList?: number;
}

export interface AttributeValues {
  object?: boolean;
  writable?: boolean;
  value?: [string | number | boolean, string?];
  notification?: number;
  accessList?: string[];
}

export interface DeviceData {
  paths: PathSet;
  timestamps: VersionedMap<Path, number>;
  attributes: VersionedMap<Path, Attributes>;
  trackers: Map<Path, { [name: string]: number }>;
  changes: Set<string>;
}

export type VirtualParameterDeclaration = [
  Path,
  {
    path?: number;
    object?: number;
    writable?: number;
    value?: number;
    notification?: number;
    accessList?: number;
  }?,
  {
    path?: [number, number];
    object?: boolean;
    writable?: boolean;
    value?: [string | number | boolean, string?];
    notification?: number;
    accessList?: string[];
  }?,
];

export interface SyncState {
  refreshAttributes: {
    exist: Set<Path>;
    object: Set<Path>;
    writable: Set<Path>;
    value: Set<Path>;
    notification: Set<Path>;
    accessList: Set<Path>;
  };
  spv: Map<Path, [string | number | boolean, string]>;
  spa: Map<Path, { notification: number | null; accessList: string[] | null }>;
  gpn: Set<Path>;
  gpnPatterns: Map<Path, number>;
  tags: Map<Path, boolean>;
  virtualParameterDeclarations: VirtualParameterDeclaration[][];
  instancesToDelete: Map<Path, Set<Path>>;
  instancesToCreate: Map<Path, InstanceSet>;
  downloadsToDelete: Set<Path>;
  downloadsToCreate: InstanceSet;
  downloadsValues: Map<Path, string | number>;
  downloadsDownload: Map<Path, number>;
  uploadsToDelete: Set<Path>;
  uploadsToCreate: InstanceSet;
  uploadsValues: Map<Path, string | number>;
  uploadsUpload: Map<Path, number>;
  reboot: number;
  factoryReset: number;
}

export interface SessionContext {
  sessionId: string;
  timestamp: number;
  deviceId: string;
  deviceData: DeviceData;
  cwmpVersion: string;
  timeout: number;
  provisions: [string, ...Value[]][];
  channels: { [channel: string]: number };
  virtualParameters: [
    string,
    AttributeTimestamps,
    AttributeValues,
    AttributeTimestamps,
    AttributeValues,
  ][][];
  revisions: number[];
  rpcCount: number;
  iteration: number;
  cycle: number;
  extensionsCache: any;
  declarations: Declaration[][];
  faults: { [channel: string]: SessionFault };
  retries: { [channel: string]: number };
  cacheSnapshot: string;
  httpResponse: ServerResponse;
  httpRequest: IncomingMessage;
  faultsTouched: { [channel: string]: boolean };
  presetCycles: number;
  new: boolean;
  debug: boolean;
  state: number;
  authState: number;
  tasks: Task[];
  operations: { [commandKey: string]: Operation };
  syncState: SyncState | undefined;
  lastActivity: number;
  extendLock: number;
  rpcRequest: AcsRequest | undefined;
  operationsTouched: { [commandKey: string]: 1 | 0 };
  provisionsRet: any[];
  doneTasks: string[];
}

export interface Task {
  _id?: string;
  name: string;
  parameterNames?: string[];
  parameterValues?: [string, string | number | boolean, string?][];
  objectName?: string;
  fileType?: string;
  fileName?: string;
  targetFileName?: string;
  expiry?: number;
  provisions?: [string, ...Value[]][];
}

export interface Operation {
  name: string;
  timestamp: number;
  provisions: [string, ...Value[]][];
  channels: { [channel: string]: number };
  retries: { [channel: string]: number };
  args: {
    instance: string;
    fileType: string;
    fileName: string;
    targetFileName?: string;
  };
}

export type AcsRequest =
  | GetParameterNames
  | GetParameterValues
  | GetParameterAttributes
  | SetParameterValues
  | SetParameterAttributes
  | AddObject
  | DeleteObject
  | FactoryReset
  | Reboot
  | Download
  | Upload;

export interface GetParameterNames {
  name: "GetParameterNames";
  parameterPath: string;
  nextLevel: boolean;
}

export interface GetParameterValues {
  name: "GetParameterValues";
  parameterNames: string[];
}

export interface GetParameterAttributes {
  name: "GetParameterAttributes";
  parameterNames: string[];
}

export interface SetParameterValues {
  name: "SetParameterValues";
  parameterList: [string, boolean | number | string, string][];
  parameterKey?: string;
  DATETIME_MILLISECONDS?: boolean;
  BOOLEAN_LITERAL?: boolean;
}

export interface SetParameterAttributes {
  name: "SetParameterAttributes";
  parameterList: [string, number | null, string[] | null][];
}

export interface AddObject {
  name: "AddObject";
  objectName: string;
  parameterKey?: string;
  instanceValues: Record<string, string>;
}

export interface DeleteObject {
  name: "DeleteObject";
  objectName: string;
  parameterKey?: string;
}

export interface FactoryReset {
  name: "FactoryReset";
  commandKey?: string;
}

export interface Reboot {
  name: "Reboot";
  commandKey?: string;
}

export interface Download {
  name: "Download";
  commandKey: string;
  instance: string;
  fileType: string;
  fileName?: string;
  url?: string;
  username?: string;
  password?: string;
  fileSize?: number;
  targetFileName?: string;
  delaySeconds?: number;
  successUrl?: string;
  failureUrl?: string;
}

export interface Upload {
  name: "Upload";
  commandKey: string;
  instance: string;
  fileType: string;
  fileName?: string;
  url?: string;
  username?: string;
  password?: string;
  delaySeconds?: number;
}
// ...
export interface Declaration {
  path: Path;
  pathGet: number;
  pathSet: number | [number, number] | undefined;
  attrGet:
    | {
        object?: number;
        writable?: number;
        value?: number;
        notification?: number;
        accessList?: number;
      }
    | undefined;
  attrSet:
    | {
        object?: boolean;
        writable?: boolean;
        value?: [string | number | boolean, string?];
        notification?: number;
        accessList?: string[];
      }
    | undefined;
  defer: boolean;
}